fix(consent): address review — em-dash, move to components/, responsive toast
Per ryanbarlow97's review:
- en.json: drop the em-dash from the consent error string.
- Move MarketingConsentToast into src/client/components/ (fix imports,
  registration, and test path accordingly).
- Toast is now a full-width top banner on small screens and the top-right
  236px card on sm+ — fixes the cramped/word-cutoff mobile layout (the old
  wrapping button label) while keeping the top-right placement on desktop.
